untangled the React fast-refresh lint failure, addressed the Cypress PR feedback, and verified everything end-to-end.
