+++ frontend/src/contexts/FeatureFlagsContext.tsx
@@ -1,14 +1,7 @@
-import React, { createContext, useCallback, useContext, useEffect, useMemo, useState } from 'react';
+import React, { useCallback, useEffect, useMemo, useState } from 'react';
 import { featureFlagsAPI } from '../services/api';
-
-interface FeatureFlagsContextValue {
-  flags: Record<string, boolean>;
-  loading: boolean;
-  refresh: () => Promise<void>;
-  isEnabled: (flag: string) => boolean;
-}
-
-const FeatureFlagsContext = createContext<FeatureFlagsContextValue | undefined>(undefined);
+import { FeatureFlagsContext } from './FeatureFlagsContextData';
+import type { FeatureFlagsContextValue } from './FeatureFlagsContextData';
 
 export const FeatureFlagsProvider: React.FC<{ children: React.ReactNode }> = ({ children }) => {
   const [flags, setFlags] = useState<Record<string, boolean>>({});
@@ -42,12 +35,4 @@ export const FeatureFlagsProvider: React.FC<{ children: React.ReactNode }> = ({ 
   );
 
   return <FeatureFlagsContext.Provider value={value}>{children}</FeatureFlagsContext.Provider>;
-};
-
-export const useFeatureFlags = (): FeatureFlagsContextValue => {
-  const context = useContext(FeatureFlagsContext);
-  if (!context) {
-    throw new Error('useFeatureFlags must be used within a FeatureFlagsProvider');
-  }
-  return context;
 };
